Extract Actor Actions, Reactions, and Animations into their own folders
