+++ app/main.js
@@ -110,6 +110,30 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _act
 
 /***/ }),
 
+/***/ "./src/avo/actions/index.js":
+/*!**********************************!*\
+  !*** ./src/avo/actions/index.js ***!
+  \**********************************/
+/*! exports provided: STANDARD_ACTIONS */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"STANDARD_ACTIONS\", function() { return STANDARD_ACTIONS; });\n/* harmony import */ var _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/misc/constants */ \"./src/avo/misc/constants.js\");\n/* harmony import */ var _avo_story_elements_particle__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/story-elements/particle */ \"./src/avo/story-elements/particle.js\");\n\n\nvar STANDARD_ACTIONS = {\n  IDLE: {\n    type: _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].IDLE,\n    steps: 1,\n    script: function script(app, actor, action, actionAttr, step) {\n      actor.animationName = 'idle';\n    }\n  },\n  MOVE: {\n    type: _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].CONTINUOUS,\n    steps: 6 * 2,\n    script: function script(app, actor, action, actionAttr, step) {\n      var acceleration = actor.stats.acceleration || 0;\n      var actionRotation = Math.atan2(actionAttr.y, actionAttr.x);\n      var moveX = actor.moveX + acceleration * Math.cos(actionRotation);\n      var moveY = actor.moveY + acceleration * Math.sin(actionRotation);\n\n      if (actor.stats.maxSpeed >= 0) {\n        var maxSpeed = actor.stats.maxSpeed;\n        var correctedSpeed = Math.min(maxSpeed, Math.sqrt(moveX * moveX + moveY * moveY));\n        var moveRotation = Math.atan2(moveY, moveX);\n        moveX = correctedSpeed * Math.cos(moveRotation);\n        moveY = correctedSpeed * Math.sin(moveRotation);\n      }\n\n      actor.moveX = moveX;\n      actor.moveY = moveY;\n      actor.rotation = actionRotation;\n      if (0 * 2 <= step && step < 1 * 2) actor.animationName = 'move-1';else if (1 * 2 <= step && step < 3 * 2) actor.animationName = 'move-2';else if (3 * 2 <= step && step < 4 * 2) actor.animationName = 'move-1';else if (4 * 2 <= step && step < 6 * 2) actor.animationName = 'move-3';\n    }\n  },\n  ATTACK: {\n    type: _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].STANDARD,\n    steps: 15,\n    script: function script(app, actor, action, actionAttr, step) {\n      if (step < 10) {\n        actor.animationName = 'attack-windup';\n      } else if (step === 10) {\n        var particle = new _avo_story_elements_particle__WEBPACK_IMPORTED_MODULE_1__[\"default\"](app, {\n          x: actor.x + Math.cos(actor.rotation) * actor.size * 0.8,\n          y: actor.y + Math.sin(actor.rotation) * actor.size * 0.8,\n          size: actor.size * 1,\n          duration: 5,\n          source: actor,\n          ignoreSource: true,\n          stats: {\n            attackPower: 20,\n            pushPower: 8,\n            pushAngle: actor.rotation,\n            pushDuration: 6,\n            pushDecay: 1\n          },\n          scripts: {\n            'collision': function collision(app, particle, target) {\n              if (target && target.stats) {\n                target.stats.health = Math.max((target.stats.health || 0) - particle.stats.attackPower, 0);\n                particle.applyEffect({\n                  name: 'push',\n                  attr: {\n                    power: particle.stats.pushPower,\n                    angle: particle.stats.pushAngle,\n                    decay: particle.stats.pushDecay\n                  },\n                  duration: particle.stats.pushDuration,\n                  stacking: _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"EFFECTS_STACKING\"].STACK\n                }, target);\n              }\n            }\n          },\n          animationScript: function animationScript(app, element, canvas) {\n            var options = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : {};\n            var camera = app.camera;\n            var layer = options.layer || '';\n            if (!canvas) return; // Simple shadow\n\n            canvas.fillStyle = 'rgba(238, 238, 204, 0.5)';\n            canvas.beginPath();\n            canvas.arc(element.x + camera.x, element.y + camera.y, element.size / 2, 0, 2 * Math.PI);\n            canvas.fill();\n          }\n        });\n        app.particles.push(particle);\n        actor.animationName = 'attack-active';\n      } else {\n        actor.animationName = 'attack-winddown';\n      }\n    }\n  }\n};\n\n//# sourceURL=webpack:///./src/avo/actions/index.js?");
+
+/***/ }),
+
+/***/ "./src/avo/animations/index.js":
+/*!*************************************!*\
+  !*** ./src/avo/animations/index.js ***!
+  \*************************************/
+/*! exports provided: STANDARD_ANIMATIONS */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"STANDARD_ANIMATIONS\", function() { return STANDARD_ANIMATIONS; });\n/* harmony import */ var _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/misc/constants */ \"./src/avo/misc/constants.js\");\n\nvar STANDARD_ANIMATIONS = {\n  ACTOR: function ACTOR(app, element, canvas) {\n    var options = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : {};\n    var camera = app.camera;\n    var layer = options.layer || '';\n    if (!canvas) return; // Simple shadow\n\n    canvas.fillStyle = 'rgba(0, 0, 0, 0.5)'; // --------\n    // Temporary 'animation'\n\n    if (element.animationName === 'idle') canvas.fillStyle = 'rgba(0, 0, 0, 0.5)';else if (element.animationName === 'move-1') canvas.fillStyle = 'rgba(0, 128, 128, 0.5)';else if (element.animationName === 'move-2') canvas.fillStyle = 'rgba(0, 160, 128, 0.5)';else if (element.animationName === 'move-3') canvas.fillStyle = 'rgba(0, 128, 160, 0.5)';else if (element.animationName === 'attack-windup') canvas.fillStyle = 'rgba(192, 192, 0, 0.5)';else if (element.animationName === 'attack-active') canvas.fillStyle = 'rgba(255, 0, 0, 0.5)';else if (element.animationName === 'attack-winddown') canvas.fillStyle = 'rgba(192, 128, 0, 0.5)'; // --------\n\n    canvas.beginPath();\n    canvas.arc(element.x + camera.x, element.y + camera.y, element.size / 2, 0, 2 * Math.PI);\n    canvas.fill(); // Simple direction\n    // --------\n\n    canvas.strokeStyle = 'rgba(0, 0, 0, 0.5)';\n    canvas.lineWidth = 2;\n    canvas.beginPath();\n    canvas.moveTo(element.x, element.y);\n    canvas.lineTo(element.x + Math.cos(element.rotation) * element.size * 0.6, element.y + Math.sin(element.rotation) * element.size * 0.6);\n    canvas.stroke(); // --------\n    // Paint actor sprite\n    // --------\n\n    if (this.animationSpritesheet) {\n      var SPRITE_SIZE = 48;\n      var SPRITE_OFFSET_X = 0;\n      var SPRITE_OFFSET_Y = -8;\n      var srcSizeX = SPRITE_SIZE;\n      var srcSizeY = SPRITE_SIZE;\n      var srcX = 0;\n      var srcY = 0;\n\n      switch (element.direction) {\n        case _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"DIRECTIONS\"].SOUTH:\n          srcX = SPRITE_SIZE * 0;\n          break;\n\n        case _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"DIRECTIONS\"].NORTH:\n          srcX = SPRITE_SIZE * 1;\n          break;\n\n        case _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"DIRECTIONS\"].EAST:\n          srcX = SPRITE_SIZE * 2;\n          break;\n\n        case _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"DIRECTIONS\"].WEST:\n          srcX = SPRITE_SIZE * 3;\n          break;\n      }\n\n      switch (element.animationName) {\n        case 'move-1':\n          srcY = SPRITE_SIZE * 1;\n          break;\n\n        case 'move-2':\n          srcY = SPRITE_SIZE * 2;\n          break;\n\n        case 'move-3':\n          srcY = SPRITE_SIZE * 3;\n          break;\n\n        case 'attack-windup':\n          srcY = SPRITE_SIZE * 4;\n          break;\n\n        case 'attack-active':\n          srcY = SPRITE_SIZE * 5;\n          break;\n\n        case 'attack-winddown':\n          srcY = SPRITE_SIZE * 5;\n          break;\n      }\n\n      var tgtSizeX = SPRITE_SIZE;\n      var tgtSizeY = SPRITE_SIZE;\n      var tgtX = Math.floor(element.x - srcSizeX / 2 + SPRITE_OFFSET_X);\n      var tgtY = Math.floor(element.y - srcSizeY / 2 + SPRITE_OFFSET_Y);\n      canvas.drawImage(this.animationSpritesheet.img, srcX, srcY, srcSizeX, srcSizeY, tgtX, tgtY, tgtSizeX, tgtSizeY);\n    } // --------\n    // Paint UI elements\n    // --------\n\n\n    var healthOffsetY = 7;\n    var healthRatio = element.stats.maxHealth > 0 ? (element.stats.health || 0) / element.stats.maxHealth : 0;\n    canvas.strokeStyle = 'rgba(0, 0, 0)';\n    canvas.lineWidth = 4;\n    canvas.beginPath();\n    canvas.moveTo(element.x - element.size / 3, element.y + element.size / 2 + healthOffsetY);\n    canvas.lineTo(element.x + element.size / 3, element.y + element.size / 2 + healthOffsetY);\n    canvas.stroke();\n    canvas.strokeStyle = 'rgba(255, 0, 0)';\n    canvas.lineWidth = 2;\n    canvas.beginPath();\n    canvas.moveTo(element.x - element.size / 3 * healthRatio, element.y + element.size / 2 + healthOffsetY);\n    canvas.lineTo(element.x + element.size / 3 * healthRatio, element.y + element.size / 2 + healthOffsetY);\n    canvas.stroke();\n    healthOffsetY = 4;\n    canvas.font = '8px Arial';\n    canvas.fillStyle = 'rgba(204, 68, 68)';\n    canvas.textBaseline = 'hanging';\n    canvas.textAlign = 'right';\n    canvas.fillText('❤️', element.x - element.size / 3, element.y + element.size / 2 + healthOffsetY);\n    canvas.textAlign = 'left';\n    canvas.fillText(Math.floor(element.stats.health), element.x + element.size / 3, element.y + element.size / 2 + healthOffsetY); // --------\n  }\n};\n\n//# sourceURL=webpack:///./src/avo/animations/index.js?");
+
+/***/ }),
+
 /***/ "./src/avo/avo-adventure.js":
 /*!**********************************!*\
   !*** ./src/avo/avo-adventure.js ***!
@@ -170,6 +194,18 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 
 /***/ }),
 
+/***/ "./src/avo/reactions/index.js":
+/*!************************************!*\
+  !*** ./src/avo/reactions/index.js ***!
+  \************************************/
+/*! exports provided: STANDARD_REACTIONS */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"STANDARD_REACTIONS\", function() { return STANDARD_REACTIONS; });\nvar STANDARD_REACTIONS = {\n  DAMAGE: function DAMAGE(app, actor, effect) {},\n  PUSH: function PUSH(app, actor, effect) {\n    var power = effect.attr && effect.attr.power || 0;\n    var angle = effect.attr && effect.attr.angle || 0;\n    actor.pushX += power * Math.cos(angle);\n    actor.pushY += power * Math.sin(angle);\n\n    if (effect.attr.decay && effect.attr.power) {\n      effect.attr.power = Math.max(effect.attr.power - effect.attr.decay, 0);\n    }\n  }\n};\n\n//# sourceURL=webpack:///./src/avo/reactions/index.js?");
+
+/***/ }),
+
 /***/ "./src/avo/story-elements/actor.js":
 /*!*****************************************!*\
   !*** ./src/avo/story-elements/actor.js ***!
@@ -178,7 +214,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/misc/constants */ \"./src/avo/misc/constants.js\");\n/* harmony import */ var _story_element__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./story-element */ \"./src/avo/story-elements/story-element.js\");\n/* harmony import */ var _particle__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./particle */ \"./src/avo/story-elements/particle.js\");\nfunction _typeof(obj) { if (typeof Symbol === \"function\" && typeof Symbol.iterator === \"symbol\") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === \"function\" && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }; } return _typeof(obj); }\n\nfunction ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); keys.push.apply(keys, symbols); } return keys; }\n\nfunction _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; if (i % 2) { ownKeys(source, true).forEach(function (key) { _defineProperty(target, key, source[key]); }); } else if (Object.getOwnPropertyDescriptors) { Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)); } else { ownKeys(source).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } } return target; }\n\nfunction _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } return _assertThisInitialized(self); }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\n\n\n\n\nvar Actor =\n/*#__PURE__*/\nfunction (_StoryElement) {\n  _inherits(Actor, _StoryElement);\n\n  function Actor(app) {\n    var _this;\n\n    var initialValues = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : {};\n\n    _classCallCheck(this, Actor);\n\n    _this = _possibleConstructorReturn(this, _getPrototypeOf(Actor).call(this, app));\n    _this.shape = _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"SHAPES\"].CIRCLE;\n    _this.solid = true;\n    _this.movable = true;\n    _this.stats = {\n      health: 100,\n      // TEMP\n      maxHealth: 100,\n      // TEMP\n      acceleration: 1,\n      deceleration: 1,\n      maxSpeed: 8\n    };\n    _this.intent = undefined;\n    _this.actionName = 'idle';\n    _this.actionAttr = {};\n    _this.actionStep = 0;\n    _this.actions = {\n      'idle': {\n        type: _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].IDLE,\n        steps: 1,\n        script: function script(app, actor, action, actionAttr, step) {\n          actor.animationName = 'idle';\n        }\n      },\n      'move': {\n        type: _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].CONTINUOUS,\n        steps: 6 * 2,\n        script: function script(app, actor, action, actionAttr, step) {\n          var acceleration = actor.stats.acceleration || 0;\n          var actionRotation = Math.atan2(actionAttr.y, actionAttr.x);\n          var moveX = actor.moveX + acceleration * Math.cos(actionRotation);\n          var moveY = actor.moveY + acceleration * Math.sin(actionRotation);\n\n          if (actor.stats.maxSpeed >= 0) {\n            var maxSpeed = actor.stats.maxSpeed;\n            var correctedSpeed = Math.min(maxSpeed, Math.sqrt(moveX * moveX + moveY * moveY));\n            var moveRotation = Math.atan2(moveY, moveX);\n            moveX = correctedSpeed * Math.cos(moveRotation);\n            moveY = correctedSpeed * Math.sin(moveRotation);\n          }\n\n          actor.moveX = moveX;\n          actor.moveY = moveY;\n          actor.rotation = actionRotation;\n          if (0 * 2 <= step && step < 1 * 2) actor.animationName = 'move-1';else if (1 * 2 <= step && step < 3 * 2) actor.animationName = 'move-2';else if (3 * 2 <= step && step < 4 * 2) actor.animationName = 'move-1';else if (4 * 2 <= step && step < 6 * 2) actor.animationName = 'move-3';\n        }\n      },\n      'attack': {\n        type: _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].STANDARD,\n        steps: 15,\n        script: function script(app, actor, action, actionAttr, step) {\n          if (step < 10) {\n            actor.animationName = 'attack-windup';\n          } else if (step === 10) {\n            var particle = new _particle__WEBPACK_IMPORTED_MODULE_2__[\"default\"](app, {\n              x: actor.x + Math.cos(actor.rotation) * actor.size * 0.8,\n              y: actor.y + Math.sin(actor.rotation) * actor.size * 0.8,\n              size: actor.size * 1,\n              duration: 5,\n              source: actor,\n              ignoreSource: true,\n              stats: {\n                attackPower: 20,\n                pushPower: 8,\n                pushAngle: actor.rotation,\n                pushDuration: 6,\n                pushDecay: 1\n              },\n              scripts: {\n                'collision': function collision(app, particle, target) {\n                  if (target && target.stats) {\n                    target.stats.health = Math.max((target.stats.health || 0) - particle.stats.attackPower, 0);\n                    particle.applyEffect({\n                      name: 'push',\n                      attr: {\n                        power: particle.stats.pushPower,\n                        angle: particle.stats.pushAngle,\n                        decay: particle.stats.pushDecay\n                      },\n                      duration: particle.stats.pushDuration,\n                      stacking: _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"EFFECTS_STACKING\"].STACK\n                    }, target);\n                  }\n                }\n              },\n              animationScript: function animationScript(app, element, canvas) {\n                var options = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : {};\n                var camera = app.camera;\n                var layer = options.layer || '';\n                if (!canvas) return; // Simple shadow\n\n                canvas.fillStyle = 'rgba(238, 238, 204, 0.5)';\n                canvas.beginPath();\n                canvas.arc(element.x + camera.x, element.y + camera.y, element.size / 2, 0, 2 * Math.PI);\n                canvas.fill();\n              }\n            });\n            app.particles.push(particle);\n            actor.animationName = 'attack-active';\n          } else {\n            actor.animationName = 'attack-winddown';\n          }\n        }\n      }\n    };\n    _this.scripts = {\n      'always': function always(app, actor) {}\n    };\n    _this.reactions = {\n      'damage': function damage(app, actor, effect) {},\n      'push': function push(app, actor, effect) {\n        var power = effect.attr && effect.attr.power || 0;\n        var angle = effect.attr && effect.attr.angle || 0;\n        actor.pushX += power * Math.cos(angle);\n        actor.pushY += power * Math.sin(angle);\n\n        if (effect.attr.decay && effect.attr.power) {\n          effect.attr.power = Math.max(effect.attr.power - effect.attr.decay, 0);\n        }\n      }\n    };\n\n    _this.animationScript = function (app, element, canvas) {\n      var options = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : {};\n      var camera = app.camera;\n      var layer = options.layer || '';\n      if (!canvas) return; // Simple shadow\n\n      canvas.fillStyle = 'rgba(0, 0, 0, 0.5)'; // --------\n      // Temporary 'animation'\n\n      if (element.animationName === 'idle') canvas.fillStyle = 'rgba(0, 0, 0, 0.5)';else if (element.animationName === 'move-1') canvas.fillStyle = 'rgba(0, 128, 128, 0.5)';else if (element.animationName === 'move-2') canvas.fillStyle = 'rgba(0, 160, 128, 0.5)';else if (element.animationName === 'move-3') canvas.fillStyle = 'rgba(0, 128, 160, 0.5)';else if (element.animationName === 'attack-windup') canvas.fillStyle = 'rgba(192, 192, 0, 0.5)';else if (element.animationName === 'attack-active') canvas.fillStyle = 'rgba(255, 0, 0, 0.5)';else if (element.animationName === 'attack-winddown') canvas.fillStyle = 'rgba(192, 128, 0, 0.5)'; // --------\n\n      canvas.beginPath();\n      canvas.arc(element.x + camera.x, element.y + camera.y, element.size / 2, 0, 2 * Math.PI);\n      canvas.fill(); // Simple direction\n      // --------\n\n      canvas.strokeStyle = 'rgba(0, 0, 0, 0.5)';\n      canvas.lineWidth = 2;\n      canvas.beginPath();\n      canvas.moveTo(element.x, element.y);\n      canvas.lineTo(element.x + Math.cos(element.rotation) * element.size * 0.6, element.y + Math.sin(element.rotation) * element.size * 0.6);\n      canvas.stroke(); // --------\n      // Paint actor sprite\n      // --------\n\n      if (this.animationSpritesheet) {\n        var SPRITE_SIZE = 48;\n        var SPRITE_OFFSET_X = 0;\n        var SPRITE_OFFSET_Y = -8;\n        var srcSizeX = SPRITE_SIZE;\n        var srcSizeY = SPRITE_SIZE;\n        var srcX = 0;\n        var srcY = 0;\n\n        switch (element.direction) {\n          case _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"DIRECTIONS\"].SOUTH:\n            srcX = SPRITE_SIZE * 0;\n            break;\n\n          case _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"DIRECTIONS\"].NORTH:\n            srcX = SPRITE_SIZE * 1;\n            break;\n\n          case _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"DIRECTIONS\"].EAST:\n            srcX = SPRITE_SIZE * 2;\n            break;\n\n          case _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"DIRECTIONS\"].WEST:\n            srcX = SPRITE_SIZE * 3;\n            break;\n        }\n\n        switch (element.animationName) {\n          case 'move-1':\n            srcY = SPRITE_SIZE * 1;\n            break;\n\n          case 'move-2':\n            srcY = SPRITE_SIZE * 2;\n            break;\n\n          case 'move-3':\n            srcY = SPRITE_SIZE * 3;\n            break;\n\n          case 'attack-windup':\n            srcY = SPRITE_SIZE * 4;\n            break;\n\n          case 'attack-active':\n            srcY = SPRITE_SIZE * 5;\n            break;\n\n          case 'attack-winddown':\n            srcY = SPRITE_SIZE * 5;\n            break;\n        }\n\n        var tgtSizeX = SPRITE_SIZE;\n        var tgtSizeY = SPRITE_SIZE;\n        var tgtX = Math.floor(element.x - srcSizeX / 2 + SPRITE_OFFSET_X);\n        var tgtY = Math.floor(element.y - srcSizeY / 2 + SPRITE_OFFSET_Y);\n        canvas.drawImage(this.animationSpritesheet.img, srcX, srcY, srcSizeX, srcSizeY, tgtX, tgtY, tgtSizeX, tgtSizeY);\n      } // --------\n      // Paint UI elements\n      // --------\n\n\n      var healthOffsetY = 7;\n      var healthRatio = element.stats.maxHealth > 0 ? (element.stats.health || 0) / element.stats.maxHealth : 0;\n      canvas.strokeStyle = 'rgba(0, 0, 0)';\n      canvas.lineWidth = 4;\n      canvas.beginPath();\n      canvas.moveTo(element.x - element.size / 3, element.y + element.size / 2 + healthOffsetY);\n      canvas.lineTo(element.x + element.size / 3, element.y + element.size / 2 + healthOffsetY);\n      canvas.stroke();\n      canvas.strokeStyle = 'rgba(255, 0, 0)';\n      canvas.lineWidth = 2;\n      canvas.beginPath();\n      canvas.moveTo(element.x - element.size / 3 * healthRatio, element.y + element.size / 2 + healthOffsetY);\n      canvas.lineTo(element.x + element.size / 3 * healthRatio, element.y + element.size / 2 + healthOffsetY);\n      canvas.stroke();\n      healthOffsetY = 4;\n      canvas.font = '8px Arial';\n      canvas.fillStyle = 'rgba(204, 68, 68)';\n      canvas.textBaseline = 'hanging';\n      canvas.textAlign = 'right';\n      canvas.fillText('❤️', element.x - element.size / 3, element.y + element.size / 2 + healthOffsetY);\n      canvas.textAlign = 'left';\n      canvas.fillText(Math.floor(element.stats.health), element.x + element.size / 3, element.y + element.size / 2 + healthOffsetY); // --------\n    }; // Set initial values\n\n\n    Object.assign(_assertThisInitialized(_this), initialValues);\n    return _this;\n  }\n\n  _createClass(Actor, [{\n    key: \"play\",\n    value: function play() {\n      var app = this._app; // Run script: \"always execute on every frame\"\n\n      this.scripts.always && this.scripts.always(app, this); // TODO: copy processEffects to Particles, too.\n\n      this.processEffects();\n      this.processIntent();\n      this.processActions(); // TODO // TEMP - move this into this.scripts.always() ?\n\n      if (this.stats.health <= 0) {\n        this._expired = true;\n      }\n\n      if (this.stats.health < this.stats.maxHealth) {\n        this.stats.health += 0.05;\n      } // Upkeep: deceleration\n\n\n      if (this.actionName !== 'move') {\n        var deceleration = this.stats.deceleration || 0;\n        var curRotation = Math.atan2(this.moveY, this.moveX);\n        var curMoveSpeed = Math.sqrt(this.moveX * this.moveX + this.moveY * this.moveY);\n        var newMoveSpeed = Math.max(0, curMoveSpeed - deceleration);\n        this.moveX = newMoveSpeed * Math.cos(curRotation);\n        this.moveY = newMoveSpeed * Math.sin(curRotation);\n      }\n    }\n  }, {\n    key: \"processIntent\",\n    value: function processIntent() {\n      // Translate intent into action.\n      var action = this.actions[this.actionName];\n\n      if (!action) {\n        // Sanity check: if the Actor isn't doing anything, go idle.\n        this.goIdle(); // This is just a failsafe - the actor should ALWAYS have an action, even if it's the IDLE action.\n      } else if (!this.intent) {\n        // If the Actor has no intent (e.g. player has no key input)...\n        // ...cancel any currently cancellable actions.\n        // (Obviously, ignore this if the \n        if (action.type === _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].CONTINUOUS) this.goIdle();\n      } else {\n        // Actor intends to perform a new action.\n        // First, can the current action be overwritten by a new action?\n        if (action.type === _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].IDLE || action.type === _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].CONTINUOUS) {\n          // Second, check if the new action is different from the old one. \n          // Reset the actionStep counter if that's the case.\n          if (this.actionName !== this.intent.name) {\n            this.actionStep = 0;\n          } // Finally, convert the intent into the new action.\n\n\n          this.actionName = this.intent.name;\n          this.actionAttr = this.intent.attr ? _objectSpread({}, this.intent.attr) : {};\n        }\n      }\n    }\n  }, {\n    key: \"goIdle\",\n    value: function goIdle() {\n      this.actionName = 'idle';\n      this.actionStep = 0;\n      this.actionAttr = {};\n    }\n  }, {\n    key: \"processEffects\",\n    value: function processEffects() {\n      var _this2 = this;\n\n      var app = this._app;\n      this.effects.forEach(function (effect) {\n        var script = _this2.reactions[effect.name];\n        script && script(app, _this2, effect);\n        effect.duration--;\n      });\n      this.effects = this.effects.filter(function (effect) {\n        return effect.duration > 0;\n      });\n    }\n  }, {\n    key: \"processActions\",\n    value: function processActions() {\n      var app = this._app;\n      var action = this.actions[this.actionName];\n      if (!action) return;\n      action.script(app, this, action, this.actionAttr, this.actionStep);\n      this.actionStep += 1;\n\n      if (this.actionStep >= action.steps) {\n        // Is the action over?\n        this.actionStep = 0; // If it's over (and doesn't loop), revert to default.\n\n        if (action.type === _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].STANDARD || action.type === _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].SPECIAL_ONCE) {\n          this.goIdle();\n        }\n      }\n    }\n  }]);\n\n  return Actor;\n}(_story_element__WEBPACK_IMPORTED_MODULE_1__[\"default\"]);\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Actor);\n\n//# sourceURL=webpack:///./src/avo/story-elements/actor.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/misc/constants */ \"./src/avo/misc/constants.js\");\n/* harmony import */ var _avo_actions__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/actions */ \"./src/avo/actions/index.js\");\n/* harmony import */ var _avo_reactions__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/reactions */ \"./src/avo/reactions/index.js\");\n/* harmony import */ var _avo_animations__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! @avo/animations */ \"./src/avo/animations/index.js\");\n/* harmony import */ var _story_element__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./story-element */ \"./src/avo/story-elements/story-element.js\");\n/* harmony import */ var _particle__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./particle */ \"./src/avo/story-elements/particle.js\");\nfunction _typeof(obj) { if (typeof Symbol === \"function\" && typeof Symbol.iterator === \"symbol\") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === \"function\" && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }; } return _typeof(obj); }\n\nfunction ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); keys.push.apply(keys, symbols); } return keys; }\n\nfunction _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; if (i % 2) { ownKeys(source, true).forEach(function (key) { _defineProperty(target, key, source[key]); }); } else if (Object.getOwnPropertyDescriptors) { Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)); } else { ownKeys(source).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } } return target; }\n\nfunction _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } return _assertThisInitialized(self); }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\n\n\n\n\n\n\n\nvar Actor =\n/*#__PURE__*/\nfunction (_StoryElement) {\n  _inherits(Actor, _StoryElement);\n\n  function Actor(app) {\n    var _this;\n\n    var initialValues = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : {};\n\n    _classCallCheck(this, Actor);\n\n    _this = _possibleConstructorReturn(this, _getPrototypeOf(Actor).call(this, app));\n    _this.shape = _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"SHAPES\"].CIRCLE;\n    _this.solid = true;\n    _this.movable = true;\n    _this.stats = {\n      health: 100,\n      // TEMP\n      maxHealth: 100,\n      // TEMP\n      acceleration: 1,\n      deceleration: 1,\n      maxSpeed: 8\n    };\n    _this.intent = undefined;\n    _this.actionName = 'idle';\n    _this.actionAttr = {};\n    _this.actionStep = 0;\n    _this.actions = {\n      'idle': _avo_actions__WEBPACK_IMPORTED_MODULE_1__[\"STANDARD_ACTIONS\"].IDLE,\n      'move': _avo_actions__WEBPACK_IMPORTED_MODULE_1__[\"STANDARD_ACTIONS\"].MOVE,\n      'attack': _avo_actions__WEBPACK_IMPORTED_MODULE_1__[\"STANDARD_ACTIONS\"].ATTACK\n    };\n    _this.scripts = {\n      'always': function always(app, actor) {}\n    };\n    _this.reactions = {\n      'damage': _avo_reactions__WEBPACK_IMPORTED_MODULE_2__[\"STANDARD_REACTIONS\"].DAMAGE,\n      'push': _avo_reactions__WEBPACK_IMPORTED_MODULE_2__[\"STANDARD_REACTIONS\"].PUSH\n    };\n    _this.animationScript = _avo_animations__WEBPACK_IMPORTED_MODULE_3__[\"STANDARD_ANIMATIONS\"].ACTOR; // Set initial values\n\n    Object.assign(_assertThisInitialized(_this), initialValues);\n    return _this;\n  }\n\n  _createClass(Actor, [{\n    key: \"play\",\n    value: function play() {\n      var app = this._app; // Run script: \"always execute on every frame\"\n\n      this.scripts.always && this.scripts.always(app, this); // TODO: copy processEffects to Particles, too.\n\n      this.processEffects();\n      this.processIntent();\n      this.processActions(); // TODO // TEMP - move this into this.scripts.always() ?\n\n      if (this.stats.health <= 0) {\n        this._expired = true;\n      }\n\n      if (this.stats.health < this.stats.maxHealth) {\n        this.stats.health += 0.05;\n      } // Upkeep: deceleration\n\n\n      if (this.actionName !== 'move') {\n        var deceleration = this.stats.deceleration || 0;\n        var curRotation = Math.atan2(this.moveY, this.moveX);\n        var curMoveSpeed = Math.sqrt(this.moveX * this.moveX + this.moveY * this.moveY);\n        var newMoveSpeed = Math.max(0, curMoveSpeed - deceleration);\n        this.moveX = newMoveSpeed * Math.cos(curRotation);\n        this.moveY = newMoveSpeed * Math.sin(curRotation);\n      }\n    }\n  }, {\n    key: \"processIntent\",\n    value: function processIntent() {\n      // Translate intent into action.\n      var action = this.actions[this.actionName];\n\n      if (!action) {\n        // Sanity check: if the Actor isn't doing anything, go idle.\n        this.goIdle(); // This is just a failsafe - the actor should ALWAYS have an action, even if it's the IDLE action.\n      } else if (!this.intent) {\n        // If the Actor has no intent (e.g. player has no key input)...\n        // ...cancel any currently cancellable actions.\n        // (Obviously, ignore this if the \n        if (action.type === _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].CONTINUOUS) this.goIdle();\n      } else {\n        // Actor intends to perform a new action.\n        // First, can the current action be overwritten by a new action?\n        if (action.type === _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].IDLE || action.type === _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].CONTINUOUS) {\n          // Second, check if the new action is different from the old one. \n          // Reset the actionStep counter if that's the case.\n          if (this.actionName !== this.intent.name) {\n            this.actionStep = 0;\n          } // Finally, convert the intent into the new action.\n\n\n          this.actionName = this.intent.name;\n          this.actionAttr = this.intent.attr ? _objectSpread({}, this.intent.attr) : {};\n        }\n      }\n    }\n  }, {\n    key: \"goIdle\",\n    value: function goIdle() {\n      this.actionName = 'idle';\n      this.actionStep = 0;\n      this.actionAttr = {};\n    }\n  }, {\n    key: \"processEffects\",\n    value: function processEffects() {\n      var _this2 = this;\n\n      var app = this._app;\n      this.effects.forEach(function (effect) {\n        var script = _this2.reactions[effect.name];\n        script && script(app, _this2, effect);\n        effect.duration--;\n      });\n      this.effects = this.effects.filter(function (effect) {\n        return effect.duration > 0;\n      });\n    }\n  }, {\n    key: \"processActions\",\n    value: function processActions() {\n      var app = this._app;\n      var action = this.actions[this.actionName];\n      if (!action) return;\n      action.script(app, this, action, this.actionAttr, this.actionStep);\n      this.actionStep += 1;\n\n      if (this.actionStep >= action.steps) {\n        // Is the action over?\n        this.actionStep = 0; // If it's over (and doesn't loop), revert to default.\n\n        if (action.type === _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].STANDARD || action.type === _avo_misc_constants__WEBPACK_IMPORTED_MODULE_0__[\"ACTION_TYPES\"].SPECIAL_ONCE) {\n          this.goIdle();\n        }\n      }\n    }\n  }]);\n\n  return Actor;\n}(_story_element__WEBPACK_IMPORTED_MODULE_4__[\"default\"]);\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Actor);\n\n//# sourceURL=webpack:///./src/avo/story-elements/actor.js?");
 
 /***/ }),
