+++ src/avo/story-elements/actor.js
@@ -1,4 +1,7 @@
-import { ACTION_TYPES, DIRECTIONS, EFFECTS_STACKING, MODES, SHAPES } from '@avo/misc/constants';
+import { ACTION_TYPES, SHAPES } from '@avo/misc/constants';
+import { STANDARD_ACTIONS } from '@avo/actions';
+import { STANDARD_REACTIONS } from '@avo/reactions';
+import { STANDARD_ANIMATIONS } from '@avo/animations';
 import StoryElement from './story-element';
 import Particle from './particle';
 
@@ -23,109 +26,9 @@ class Actor extends StoryElement {
     this.actionAttr = {};
     this.actionStep = 0;
     this.actions = {
-      'idle': {
-        type: ACTION_TYPES.IDLE,
-        steps: 1,
-        script: function (app, actor, action, actionAttr, step) {
-          actor.animationName = 'idle';
-        }
-      },
-      'move': {
-        type: ACTION_TYPES.CONTINUOUS,
-        steps: 6 * 2,
-        script: function (app, actor, action, actionAttr, step) {
-          const acceleration = actor.stats.acceleration || 0;
-          
-          const actionRotation = Math.atan2(actionAttr.y, actionAttr.x);
-          let moveX = actor.moveX + acceleration * Math.cos(actionRotation);
-          let moveY = actor.moveY + acceleration * Math.sin(actionRotation);
-          
-          if (actor.stats.maxSpeed >= 0) {
-            const maxSpeed = actor.stats.maxSpeed;
-            const correctedSpeed = Math.min(maxSpeed, Math.sqrt(moveX * moveX + moveY * moveY));
-            const moveRotation = Math.atan2(moveY, moveX);
-            moveX = correctedSpeed * Math.cos(moveRotation);
-            moveY = correctedSpeed * Math.sin(moveRotation);
-          }
-          
-          actor.moveX = moveX;
-          actor.moveY = moveY;
-          actor.rotation = actionRotation;
-          
-          if (0 * 2 <= step && step < 1 * 2) actor.animationName = 'move-1';
-          else if (1 * 2 <= step && step < 3 * 2) actor.animationName = 'move-2';
-          else if (3 * 2 <= step && step < 4 * 2) actor.animationName = 'move-1';
-          else if (4 * 2 <= step && step < 6 * 2) actor.animationName = 'move-3';
-        },
-      },
-      'attack': {
-        type: ACTION_TYPES.STANDARD,
-        steps: 15,
-        script: function (app, actor, action, actionAttr, step) {
-          if (step < 10) {
-
-            actor.animationName = 'attack-windup';
-
-          } else if (step === 10) {
-            
-            const particle = new Particle(app, {
-              x: actor.x + Math.cos(actor.rotation) * actor.size * 0.8,
-              y: actor.y + Math.sin(actor.rotation) * actor.size * 0.8,
-              size: actor.size * 1,
-              duration: 5,
-              source: actor,
-              ignoreSource: true,
-              stats: {
-                attackPower: 20,
-                pushPower: 8,
-                pushAngle: actor.rotation,
-                pushDuration: 6,
-                pushDecay: 1,
-              },
-              scripts: {
-                'collision': function (app, particle, target) {
-                  if (target && target.stats) {
-                    target.stats.health = Math.max((target.stats.health || 0) - particle.stats.attackPower, 0);
-                    
-                    particle.applyEffect({
-                      name: 'push',
-                      attr: {
-                        power: particle.stats.pushPower,
-                        angle: particle.stats.pushAngle,
-                        decay: particle.stats.pushDecay,
-                      },
-                      duration: particle.stats.pushDuration,
-                      stacking: EFFECTS_STACKING.STACK,
-                    }, target);
-                    
-                  }
-                },
-              },
-              animationScript: function (app, element, canvas, options = {}) {
-                const camera = app.camera;
-                const layer = options.layer || '';
-
-                if (!canvas) return;
-                
-                // Simple shadow
-                canvas.fillStyle = 'rgba(238, 238, 204, 0.5)';
-                canvas.beginPath();
-                canvas.arc(element.x + camera.x, element.y + camera.y, element.size / 2, 0, 2 * Math.PI);
-                canvas.fill();
-              },
-            });
-            app.particles.push(particle);
-            
-            actor.animationName = 'attack-active';
-            
-          } else {
-            
-            actor.animationName = 'attack-winddown';
-            
-          }
-        }
-      },
-      
+      'idle': STANDARD_ACTIONS.IDLE,
+      'move': STANDARD_ACTIONS.MOVE,
+      'attack': STANDARD_ACTIONS.ATTACK,
     };
     
     this.scripts = {
@@ -133,123 +36,11 @@ class Actor extends StoryElement {
     };
     
     this.reactions = {
-      'damage': function (app, actor, effect) {},
-      'push': function (app, actor, effect) {
-        const power = effect.attr && effect.attr.power || 0;
-        const angle = effect.attr && effect.attr.angle || 0;
-        actor.pushX += power * Math.cos(angle);
-        actor.pushY += power * Math.sin(angle);
-        
-        if (effect.attr.decay && effect.attr.power) {
-          effect.attr.power = Math.max(effect.attr.power - effect.attr.decay, 0);
-        }
-      },
+      'damage': STANDARD_REACTIONS.DAMAGE,
+      'push': STANDARD_REACTIONS.PUSH,
     };
     
-    this.animationScript = function (app, element, canvas, options = {}) {
-      
-      const camera = app.camera;
-      const layer = options.layer || '';
-
-      if (!canvas) return;
-
-      // Simple shadow
-      canvas.fillStyle = 'rgba(0, 0, 0, 0.5)';
-      // --------
-      // Temporary 'animation'
-      if (element.animationName === 'idle') canvas.fillStyle = 'rgba(0, 0, 0, 0.5)';
-      else if (element.animationName === 'move-1') canvas.fillStyle = 'rgba(0, 128, 128, 0.5)';
-      else if (element.animationName === 'move-2') canvas.fillStyle = 'rgba(0, 160, 128, 0.5)';
-      else if (element.animationName === 'move-3') canvas.fillStyle = 'rgba(0, 128, 160, 0.5)';
-      else if (element.animationName === 'attack-windup') canvas.fillStyle = 'rgba(192, 192, 0, 0.5)';
-      else if (element.animationName === 'attack-active') canvas.fillStyle = 'rgba(255, 0, 0, 0.5)';
-      else if (element.animationName === 'attack-winddown') canvas.fillStyle = 'rgba(192, 128, 0, 0.5)';
-      // --------
-      canvas.beginPath();
-      canvas.arc(element.x + camera.x, element.y + camera.y, element.size / 2, 0, 2 * Math.PI);
-      canvas.fill();
-
-      // Simple direction
-      // --------
-      canvas.strokeStyle = 'rgba(0, 0, 0, 0.5)';
-      canvas.lineWidth = 2;
-      canvas.beginPath();
-      canvas.moveTo(element.x, element.y);
-      canvas.lineTo(element.x + Math.cos(element.rotation) * element.size * 0.6,
-                      element.y + Math.sin(element.rotation) * element.size * 0.6);
-      canvas.stroke();
-      // --------
-
-      // Paint actor sprite
-      // --------
-      if (this.animationSpritesheet) {
-        const SPRITE_SIZE = 48;
-        let SPRITE_OFFSET_X = 0;
-        let SPRITE_OFFSET_Y = -8;
-
-        const srcSizeX = SPRITE_SIZE;
-        const srcSizeY = SPRITE_SIZE;
-        let srcX = 0;
-        let srcY = 0;
-        
-        switch (element.direction) {
-          case DIRECTIONS.SOUTH: srcX = SPRITE_SIZE * 0; break;
-          case DIRECTIONS.NORTH: srcX = SPRITE_SIZE * 1; break;
-          case DIRECTIONS.EAST: srcX = SPRITE_SIZE * 2; break;
-          case DIRECTIONS.WEST: srcX = SPRITE_SIZE * 3; break;
-        }
-        
-        switch (element.animationName) {
-          case 'move-1': srcY = SPRITE_SIZE * 1; break;
-          case 'move-2': srcY = SPRITE_SIZE * 2; break;
-          case 'move-3': srcY = SPRITE_SIZE * 3; break;
-          case 'attack-windup': srcY = SPRITE_SIZE * 4; break;
-          case 'attack-active': srcY = SPRITE_SIZE * 5; break;
-          case 'attack-winddown': srcY = SPRITE_SIZE * 5; break;
-        }
-
-        const tgtSizeX = SPRITE_SIZE;
-        const tgtSizeY = SPRITE_SIZE;
-        const tgtX = Math.floor(element.x - srcSizeX / 2 + SPRITE_OFFSET_X);
-        const tgtY = Math.floor(element.y - srcSizeY / 2 + SPRITE_OFFSET_Y);
-
-        canvas.drawImage(this.animationSpritesheet.img, srcX, srcY, srcSizeX, srcSizeY, tgtX, tgtY, tgtSizeX, tgtSizeY);
-      }
-      // --------
-
-      // Paint UI elements
-      // --------
-      let healthOffsetY = 7;
-      const healthRatio = (element.stats.maxHealth > 0)
-        ? (element.stats.health || 0) / element.stats.maxHealth
-        : 0;
-      canvas.strokeStyle = 'rgba(0, 0, 0)';
-      canvas.lineWidth = 4;
-      canvas.beginPath();
-      canvas.moveTo(element.x - element.size / 3,
-                    element.y + element.size / 2 + healthOffsetY);
-      canvas.lineTo(element.x + element.size / 3,
-                    element.y + element.size / 2 + healthOffsetY);
-      canvas.stroke();
-      canvas.strokeStyle = 'rgba(255, 0, 0)';
-      canvas.lineWidth = 2;
-      canvas.beginPath();
-      canvas.moveTo(element.x - (element.size / 3 * healthRatio),
-                    element.y + element.size / 2 + healthOffsetY);
-      canvas.lineTo(element.x + (element.size / 3 * healthRatio),
-                    element.y + element.size / 2 + healthOffsetY);
-      canvas.stroke();
-
-      healthOffsetY = 4;
-      canvas.font = '8px Arial';
-      canvas.fillStyle = 'rgba(204, 68, 68)';    
-      canvas.textBaseline = 'hanging';
-      canvas.textAlign = 'right';
-      canvas.fillText('❤️', element.x - element.size / 3, element.y + element.size / 2 + healthOffsetY);
-      canvas.textAlign = 'left';
-      canvas.fillText(Math.floor(element.stats.health), element.x + element.size / 3, element.y + element.size / 2 + healthOffsetY);
-      // --------
-    };
+    this.animationScript = STANDARD_ANIMATIONS.ACTOR;
     
     // Set initial values
     Object.assign(this, initialValues);
